# www/storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"Find.IT API","description":"","contact":{"name":"Rafa\u0142 Bieniek","email":"rafal@sofine.pl"},"license":{"name":"Unlicense"},"version":"1.0"},"host":"find.it:8001","basePath":"\/api","schemes":["http"],"consumes":["application\/json"],"produces":["application\/json"],"paths":{"\/category":{"get":{"tags":["Kategorie"],"summary":"Lista kategorii","responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"items":{"$ref":"#\/definitions\/Category"}},"message":{"type":"string"}}}}}}},"\/sendMessage":{"post":{"tags":["Kontakt"],"summary":"Wy\u015blij wiadomo\u015bc do uzytkownika","parameters":[{"name":"body","in":"body","description":"","required":true,"schema":{"$ref":"#\/definitions\/SendMessage"}}],"responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"items":{"$ref":"#\/definitions\/SendMessage"}},"message":{"type":"string"}}}},"404":{"description":"ResponseError","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}},"security":[{"api_key_security":[]}]}},"\/reports":{"get":{"tags":["Reports"],"summary":"Lista raportow","responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"items":{"$ref":"#\/definitions\/Report"}},"message":{"type":"string"}}}}}},"post":{"tags":["Reports"],"summary":"Utworz raport","parameters":[{"name":"body","in":"body","description":"","required":true,"schema":{"$ref":"#\/definitions\/Report"},"type":"string"}],"responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"items":{"$ref":"#\/definitions\/Report"}},"message":{"type":"string"}}}},"404":{"description":"ResponseError","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}},"security":[{"api_key_security":[]}]}},"\/reports\/{slug}":{"get":{"tags":["Reports"],"summary":"Poka\u017c raport","parameters":[{"name":"slug","in":"path","description":"","required":true,"type":"string"}],"responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"type":"object"},"message":{"type":"string"}}}}}},"put":{"tags":["Reports"],"summary":"Aktualizuj raport","parameters":[{"name":"slug","in":"path","description":"","required":true,"type":"string"},{"name":"body","in":"body","description":"","required":true,"schema":{"$ref":"#\/definitions\/Report"}}],"responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"items":{"$ref":"#\/definitions\/Report"}},"message":{"type":"string"}}}},"404":{"description":"ResponseError","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}},"security":[{"api_key_security":[]}]},"delete":{"tags":["Reports"],"summary":"Skasuj raport","parameters":[{"name":"slug","in":"path","description":"","required":true,"type":"string"}],"responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"items":{"$ref":"#\/definitions\/Report"}},"message":{"type":"string"}}}},"404":{"description":"ResponseError","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}},"security":[{"api_key_security":[]}]}},"\/search":{"post":{"tags":["Szukaj"],"summary":"Szukaj","parameters":[{"name":"body","in":"body","description":"","required":true,"schema":{"$ref":"#\/definitions\/ReportSearch"},"type":"string"}],"responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"items":{"$ref":"#\/definitions\/SendMessage"}},"message":{"type":"string"}}}},"404":{"description":"ResponseError","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}}}},"\/user\/reports":{"get":{"tags":["U\u017cytkownik"],"summary":"Lista raportow u\u017cytkownika","responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"items":{"$ref":"#\/definitions\/Report"}},"message":{"type":"string"}}}}},"security":[{"api_key_security":[]}]}},"\/auth\/login":{"post":{"tags":["Autoryzacja"],"summary":"Logowanie","parameters":[{"name":"email","in":"query","description":"The user name for login","required":true,"type":"string"},{"name":"password","in":"query","description":"The password for login in clear text","required":true,"type":"string"}],"responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"type":"object"},"message":{"type":"string"}}}},"401":{"description":"ResponseError","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}}}},"\/auth\/logout":{"get":{"tags":["Autoryzacja"],"summary":"Wylogowanie","responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"type":"object"},"message":{"type":"string"}}}},"401":{"description":"ResponseError","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}},"security":[{"api_key_security":[]}]}},"\/auth\/refresh":{"get":{"tags":["Autoryzacja"],"summary":"Reset tokena","responses":{"200":{"description":"ResponseSuccess","schema":{"properties":{"success":{"type":"boolean"},"data":{"type":"object"},"message":{"type":"string"}}}},"401":{"description":"ResponseError","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}},"security":[{"api_key_security":[]}]}}},"definitions":{"Timestamps":{"properties":{"created_at":{"description":"Creation date","type":"string","format":"date-time","example":"2017-03-01 00:00:00"},"updated_at":{"description":"Last updated","type":"string","format":"date-time","example":"2017-03-01 00:00:00"}}},"SendMessage":{"properties":{"user_id":{"description":"ID uzytkownika","type":"integer","format":"int64","example":"1"},"message":{"description":"Wiadomo\u015bc","type":"string"}}},"Category":{"required":["id","name"],"properties":{"id":{"description":"","type":"integer","example":1},"name":{"description":"","type":"string","example":"Nazwa"},"reports":{"items":{"$ref":"#\/definitions\/Report"}}}},"Report":{"required":["category_id","name","description","coordinate"],"properties":{"category_id":{"description":"","type":"integer","example":1},"name":{"description":"","type":"string","example":"Nazwa"},"description":{"description":"","type":"string","example":"opis"},"coordinate":{"properties":{"lat":{"type":"string","example":54.222222},"lng":{"type":"string","example":16.222222}},"type":"object"}}},"ReportSearch":{"required":["category","name"],"properties":{"category":{"description":"","type":"string","example":"Dokumenty"},"name":{"description":"","type":"string","example":"Kunze"}}},"Users":{"required":["name","email"],"properties":{"id":{"description":"","type":"integer","example":1},"name":{"description":"","type":"string","example":"Nazwa"},"email":{"description":"","type":"string","example":"opis"}}}},"responses":{"ResponseSuccess":{"description":"the basic response","schema":{"properties":{"success":{"type":"boolean"},"data":{"type":"object"},"message":{"type":"string"}}}},"ResponseError":{"description":"the basic response","schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string"}}}}},"securityDefinitions":{"api_key_security":{"type":"apiKey","description":"A short description for security scheme","name":"Authorization","in":"header"}}}